Move stuff from extractVideoPreviewInfo() into YoutubeStreamInfoItemExtractor and partially fix search
